Refactor: store slice 별로 분리
