+++ src/App.js
@@ -2,7 +2,7 @@ import GlobalStyle from "style/GlobalStyle";
 import MainPage from "layout/MainPage";
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
-import { tableActions } from "store";
+import { tableActions } from "store/tableSlice";
 import MOCK_DATA from "assets/MOCK_DATA.json";
 import SelectView from "components/SelectView";
