+++ src/store/index.js
@@ -1,52 +1,14 @@
-import { configureStore, createSlice } from "@reduxjs/toolkit";
-import MOCK_DATA from "assets/MOCK_DATA.json";
+import { configureStore } from "@reduxjs/toolkit";
+import currentRowReducer from "./currentRowSlice";
+import sortDataReducer from "./sortDataSlice";
+import tableReducer from "./tableSlice";
 
-const tableInitialState = {
-  tableData: [],
-};
-
-const tableSlice = createSlice({
-  name: "table",
-  initialState: tableInitialState,
-  reducers: {
-    getTableData(state, action) {
-      state.tableData = action.payload;
-    },
-  },
-});
-
-const sortDataSlice = createSlice({
-  name: "sort",
-  initialState: [...MOCK_DATA],
-  reducers: {
-    descending(state, action) {
-      state.sort((a, b) => (a[action.payload] > b[action.payload] ? -1 : 1));
-    },
-    ascending(state, action) {
-      state.sort((a, b) => (a[action.payload] > b[action.payload] ? 1 : -1));
-    },
-  },
-});
-
-const currentRowSlice = createSlice({
-  name: "currentRow",
-  initialState: { row: -1 },
-  reducers: {
-    setCurrentRow(state, action) {
-      state.row = action.payload;
-    },
-  },
-});
 const store = configureStore({
   reducer: {
-    data: tableSlice.reducer,
-    sort: sortDataSlice.reducer,
-    currentRow: currentRowSlice.reducer,
+    data: tableReducer,
+    sort: sortDataReducer,
+    currentRow: currentRowReducer,
   },
 });
-
-export const tableActions = tableSlice.actions;
-export const sortDataActions = sortDataSlice.actions;
-export const currentRowActions = currentRowSlice.actions;
 
 export default store;
